fix(logger): split ringbuf reader by platform
