+++ pkg/logger/log_reader_stub.go
@@ -1,0 +1,32 @@
+//go:build !linux
+
+/*
+ * Copyright The Kmesh Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at:
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package logger
+
+import (
+	"context"
+
+	"github.com/cilium/ebpf"
+)
+
+// StartLogReader is a no-op on non-Linux platforms because the ring buffer
+// reader depends on Linux-only eBPF APIs.
+func StartLogReader(ctx context.Context, logMapFd *ebpf.Map) {
+	_ = ctx
+	_ = logMapFd
+}
